refactor: 우승자를 찾는 로직은 board가 아닌 game이 수행하도록 변경
