+++ src/test/java/chess/domain/board/ChessBoardTest.java
@@ -1,7 +1,6 @@
 package chess.domain.board;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import chess.domain.EmptySquaresMaker;
@@ -75,29 +74,6 @@ public class ChessBoardTest {
         Score score = chessBoard.calculateScore(Color.BLACK);
 
         assertThat(score).isEqualTo(Score.of(20.5, Color.BLACK));
-    }
-
-    @DisplayName("양쪽 킹이 살아있으면 아직 승부가 나지 않는다.")
-    @Test
-    void noOneWinIfBothKingAlive() {
-        final Map<Position, Square> squares = EmptySquaresMaker.make();
-        squares.put(new Position(Rank.FIRST, File.A), King.from(Color.WHITE));
-        squares.put(new Position(Rank.FIRST, File.E), King.from(Color.BLACK));
-        ChessBoard chessBoard = new ChessBoard(squares);
-
-        assertThatThrownBy(chessBoard::findWinner).isInstanceOf(IllegalStateException.class);
-    }
-
-    @DisplayName("화이트킹이 죽으면 블랙진영의 승리이다.")
-    @Test
-    void blackWinIfWhiteDied() {
-        final Map<Position, Square> squares = EmptySquaresMaker.make();
-        squares.put(new Position(Rank.FIRST, File.E), King.from(Color.BLACK));
-        ChessBoard chessBoard = new ChessBoard(squares);
-
-        Color winner = chessBoard.findWinner();
-
-        assertThat(winner).isEqualTo(Color.BLACK);
     }
 
     @DisplayName("보드에 남아있는 Piece들을 찾는다.")
